chore: move single-use traits to relevant class. sigimm-49
