+++ src/Factories/SchemaFactory.php
@@ -7,7 +7,7 @@
  * @author Simon `Firesphere` Erkelens; Marco `Sheepy` Hermo
  * @copyright Copyright (c) 2018 - now() Firesphere & Sheepy
  * @author Signify Ltd <info@signify.co.nz>
- * Signify Ltd modified code in Nov 2024
+ * Signify Ltd modified code in July 2025
  */
 
 namespace Firesphere\SolrSearch\Factories;
@@ -16,7 +16,6 @@ use Exception;
 use Firesphere\SolrSearch\Helpers\FieldResolver;
 use Firesphere\SolrSearch\Helpers\Statics;
 use Firesphere\SolrSearch\Services\SolrCoreService;
-use Firesphere\SolrSearch\Traits\GetSetSchemaFactoryTrait;
 use SilverStripe\Control\Director;
 use SilverStripe\Core\Injector\Injector;
 use SilverStripe\Core\Manifest\ModuleLoader;
@@ -31,8 +30,6 @@ use SilverStripe\View\ViewableData;
  */
 class SchemaFactory extends ViewableData
 {
-    use GetSetSchemaFactoryTrait;
-
     /**
      * @var array Fields that always need to be stored, by Index name
      */
@@ -49,6 +46,30 @@ class SchemaFactory extends ViewableData
      * @var array Base paths to the template
      */
     protected $baseTemplatePath;
+    /**
+     * ABSOLUTE Path to template
+     *
+     * @var string
+     */
+    protected $template;
+    /**
+     * Store the value in Solr
+     *
+     * @var bool
+     */
+    protected $store = false;
+    /**
+     * Index to generate the schema for
+     *
+     * @var BaseIndex
+     */
+    protected $index;
+    /**
+     * ABSOLUTE Path to types.ss template
+     *
+     * @var string
+     */
+    protected $typesTemplate;
 
     /**
      * SchemaFactory constructor.
@@ -317,5 +338,126 @@ class SchemaFactory extends ViewableData
     public function getExtrasPath()
     {
         return $this->getTemplatePathFor('extras');
+    }
+    /**
+     * Set the store value
+     *
+     * @param bool $store
+     */
+    public function setStore(bool $store): void
+    {
+        $this->store = $store;
+    }
+
+    /**
+     * Get the Index that's being used
+     *
+     * @return BaseIndex
+     */
+    public function getIndex()
+    {
+        return $this->index;
+    }
+
+    /**
+     * Set the index that's being used and add the introspection for it
+     *
+     * @param BaseIndex $index
+     * @return SchemaFactory
+     */
+    public function setIndex($index): self
+    {
+        $this->index = $index;
+        // Add the index to the introspection as well, there's no need for a separate call here
+        // We're loading this core, why would we want the introspection from a different index?
+        $this->fieldResolver->setIndex($index);
+
+        return $this;
+    }
+
+    /**
+     * Get the name of the index being used
+     *
+     * @return string
+     */
+    public function getIndexName(): string
+    {
+        return $this->index->getIndexName();
+    }
+
+    /**
+     * Get the default field to generate df components for
+     *
+     * @return string|array
+     */
+    public function getDefaultField()
+    {
+        return $this->index->getDefaultField();
+    }
+
+    /**
+     * Get the Identifier Field for Solr
+     *
+     * @return string
+     */
+    public function getIDField(): string
+    {
+        return SolrCoreService::ID_FIELD;
+    }
+
+    /**
+     * Get the Identifier Field for Solr
+     *
+     * @return string
+     */
+    public function getClassID(): string
+    {
+        return SolrCoreService::CLASS_ID_FIELD;
+    }
+
+    /**
+     * Get the types template if defined
+     *
+     * @return string
+     */
+    public function getTypesTemplate()
+    {
+        return $this->typesTemplate;
+    }
+
+    /**
+     * Set custom types template
+     *
+     * @param string $typesTemplate
+     * @return SchemaFactory
+     */
+    public function setTypesTemplate($typesTemplate): self
+    {
+        $this->typesTemplate = $typesTemplate;
+
+        return $this;
+    }
+
+    /**
+     * Get the base template for the schema xml
+     *
+     * @return string
+     */
+    public function getTemplate()
+    {
+        return $this->template;
+    }
+
+    /**
+     * Set a custom template for schema xml
+     *
+     * @param string $template
+     * @return SchemaFactory
+     */
+    public function setTemplate($template): self
+    {
+        $this->template = $template;
+
+        return $this;
     }
 }
