+++ src/Tasks/SolrIndexTask.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Class SolrIndexTask|Firesphere\SolrSearch\Tasks\SolrIndexTask Index Solr cores
  *
@@ -6,7 +7,7 @@
  * @author Simon `Firesphere` Erkelens; Marco `Sheepy` Hermo
  * @copyright Copyright (c) 2018 - now() Firesphere & Sheepy
  * @author Signify Ltd <info@signify.co.nz>
- * Signify Ltd modified code in Aug 2024
+ * Signify Ltd modified code in July 2025
  */
 
 namespace Firesphere\SolrSearch\Tasks;
@@ -18,7 +19,6 @@ use Firesphere\SolrSearch\Indexes\BaseIndex;
 use Firesphere\SolrSearch\Services\SolrCoreService;
 use Firesphere\SolrSearch\States\SiteState;
 use Firesphere\SolrSearch\Traits\LoggerTrait;
-use Firesphere\SolrSearch\Traits\SolrIndexTrait;
 use Psr\Log\LoggerInterface;
 use ReflectionException;
 use SilverStripe\Control\Controller;
@@ -44,7 +44,6 @@ use Solarium\Exception\HttpException;
 class SolrIndexTask extends BuildTask
 {
     use LoggerTrait;
-    use SolrIndexTrait;
 
     /**
      * URLSegment of this task
@@ -70,6 +69,32 @@ class SolrIndexTask extends BuildTask
      * @var string
      */
     protected $description = 'Add or update documents to an existing Solr core.';
+    /**
+     * Debug mode enabled, default false
+     *
+     * @var bool
+     */
+    protected $debug = false;
+    /**
+     * Singleton of {@link SolrCoreService}
+     *
+     * @var SolrCoreService
+     */
+    protected $service;
+    /**
+     * @var BaseIndex Current core being indexed
+     */
+    protected $index;
+    /**
+     * @var int Number of CPU cores available
+     */
+    protected $cores = 1;
+    /**
+     * Default batch length
+     *
+     * @var int
+     */
+    protected $batchLength = 500;
 
     /**
      * SolrIndexTask constructor. Sets up the document factory
@@ -443,10 +468,112 @@ class SolrIndexTask extends BuildTask
         $this->getLogger()->error($exception->getMessage());
         $msg = sprintf(
             'Error indexing core %s on group %s,' . PHP_EOL .
-            'Please log in to the CMS to find out more about Indexing errors' . PHP_EOL,
+                'Please log in to the CMS to find out more about Indexing errors' . PHP_EOL,
             $index,
             $group
         );
         SolrLogger::logMessage('ERROR', $msg);
+    }
+
+    /**
+     * Set the {@link SolrCoreService}
+     *
+     * @param SolrCoreService $service
+     * @return self
+     */
+    public function setService(SolrCoreService $service): self
+    {
+        $this->service = $service;
+
+        return $this;
+    }
+
+    /**
+     * Is this Index in debug mode
+     *
+     * @return bool
+     */
+    public function isDebug(): bool
+    {
+        return $this->debug;
+    }
+
+    /**
+     * Set the debug mode
+     *
+     * @param bool $debug Set the task in debug mode
+     * @param bool $force Force a task in debug mode, despite e.g. being Live and not CLI
+     * @return self
+     */
+    public function setDebug(bool $debug, bool $force = false): self
+    {
+        // Make the debug a configurable, forcing it to always be false from config
+        if (!$force && SolrCoreService::config()->get('debug') === false) {
+            $debug = false;
+        }
+
+        $this->debug = $debug;
+
+        return $this;
+    }
+
+    /**
+     * Get the Index class.
+     *
+     * @return BaseIndex
+     */
+    public function getIndex(): BaseIndex
+    {
+        return $this->index;
+    }
+
+    /**
+     * Set the index class
+     *
+     * @param BaseIndex $index
+     */
+    public function setIndex(BaseIndex $index): void
+    {
+        $this->index = $index;
+    }
+
+    /**
+     * Get the amount of CPU Cores configured
+     *
+     * @return int
+     */
+    public function getCores(): int
+    {
+        return $this->cores;
+    }
+
+    /**
+     * Set the amount of CPU Cores to use
+     *
+     * @param int $cores
+     */
+    public function setCores(int $cores): void
+    {
+        $this->cores = $cores;
+    }
+
+    /**
+     * Get the length of a single batch
+     *
+     * @return int
+     */
+    public function getBatchLength(): int
+    {
+        return $this->batchLength;
+    }
+
+    /**
+     * Set the length of a single batch
+     *
+     * @param int $batchLength
+     */
+    public function setBatchLength(int $batchLength): void
+    {
+        $this->batchLength = $batchLength;
     }
 }
